backend/requirements.txt
django
djangorestframework
osmnx 
pandas
scikit-learn
psycopg2-binary